Create spread_lectures_different_days
Requires a constraint profile to have been set up already in Syllabus using the Maximum Hours to allow only 1 hour per day.  The program then creates an Equipment item for each course which has single streamed, 1 hour lectures and attaches the constraint profile to the equipment item and the equipment item to each lecture of the course.